Move GUI from Core to Editor
