+++ src/Editor/GUI/ImageButton.hpp
@@ -1,29 +1,26 @@
-#ifndef IMAGETEXTBUTTON_HPP
-#define IMAGETEXTBUTTON_HPP
+#ifndef IMAGEBUTTON_HPP
+#define IMAGEBUTTON_HPP
 
 #include "Button.hpp"
-#include "../Texture/Texture2D.hpp"
-#include "Font.hpp"
+#include <Core/Texture/Texture2D.hpp>
 
 /** @ingroup Core
  * @{
  */
 
 namespace GUI {
-    /// Button displaying both an image and text.
-    class ImageTextButton : public Button {
+    /// Button displaying an image.
+    class ImageButton : public Button {
         public:
             /// Create new button.
             /**
              * @param parent Parent widget.
              * @param texture %Button texture.
-             * @param font %Font to display text with.
-             * @param text Text to display.
              */
-            ImageTextButton(Widget* parent, Texture2D* texture, Font* font, const std::string& text);
+            ImageButton(Widget* parent, Texture2D* texture);
             
             /// Destructor.
-            virtual ~ImageTextButton();
+            virtual ~ImageButton();
             
             /// Render the widget.
             /**
@@ -32,22 +29,8 @@ namespace GUI {
              */
             void Render(int screenWidth, int screenHeight);
             
-            /// Get the image size.
-            /**
-             * @return How large the displayed image should be.
-             */
-            glm::vec2 ImageSize() const;
-            
-            /// Set the image size.
-            /**
-             * @param size How large the displayed image should be.
-             */
-            void SetImageSize(const glm::vec2& size);
-            
         private:
             Geometry::Rectangle* rectangle;
-            
-            glm::vec2 imageSize;
             
             // Shaders
             Shader* vertexShader;
@@ -57,8 +40,6 @@ namespace GUI {
             ShaderProgram* textureShaderProgram;
             
             Texture2D* texture;
-            Font* font;
-            std::string text;
     };
 }
