+++ src/Editor/GUI/ImageTextButton.hpp
@@ -1,26 +1,29 @@
-#ifndef IMAGEBUTTON_HPP
-#define IMAGEBUTTON_HPP
+#ifndef IMAGETEXTBUTTON_HPP
+#define IMAGETEXTBUTTON_HPP
 
 #include "Button.hpp"
-#include "../Texture/Texture2D.hpp"
+#include <Core/Texture/Texture2D.hpp>
+#include <Core/Font/Font.hpp>
 
 /** @ingroup Core
  * @{
  */
 
 namespace GUI {
-    /// Button displaying an image.
-    class ImageButton : public Button {
+    /// Button displaying both an image and text.
+    class ImageTextButton : public Button {
         public:
             /// Create new button.
             /**
              * @param parent Parent widget.
              * @param texture %Button texture.
+             * @param font %Font to display text with.
+             * @param text Text to display.
              */
-            ImageButton(Widget* parent, Texture2D* texture);
+            ImageTextButton(Widget* parent, Texture2D* texture, Font* font, const std::string& text);
             
             /// Destructor.
-            virtual ~ImageButton();
+            virtual ~ImageTextButton();
             
             /// Render the widget.
             /**
@@ -29,8 +32,22 @@ namespace GUI {
              */
             void Render(int screenWidth, int screenHeight);
             
+            /// Get the image size.
+            /**
+             * @return How large the displayed image should be.
+             */
+            glm::vec2 ImageSize() const;
+            
+            /// Set the image size.
+            /**
+             * @param size How large the displayed image should be.
+             */
+            void SetImageSize(const glm::vec2& size);
+            
         private:
             Geometry::Rectangle* rectangle;
+            
+            glm::vec2 imageSize;
             
             // Shaders
             Shader* vertexShader;
@@ -40,6 +57,8 @@ namespace GUI {
             ShaderProgram* textureShaderProgram;
             
             Texture2D* texture;
+            Font* font;
+            std::string text;
     };
 }
